testing refactor part 1
