+++ package.json
@@ -27,6 +27,7 @@
 		"@tiptap/extension-font-family": "^3.27.4",
 		"@tiptap/extension-text-style": "^3.27.4",
 		"@types/node": "^26.1.1",
+		"@vitest/coverage-v8": "4.1.10",
 		"file-type": "^22.0.1",
 		"jsdom": "^29.1.1",
 		"oxlint": "^1.74.0",
