+++ src/tests/unit/data/bad/json/too_big_shadow_color.json
@@ -1,0 +1,18 @@
+{
+	"content": [
+		{
+			"content": [
+				{
+					"type": "text",
+					"text": "Shadow",
+					"marks": [
+						{
+							"type": "shadowColor",
+							"attrs": { "shadowColor": "#100000000" }
+						}
+					]
+				}
+			]
+		}
+	]
+}
