+++ src/tests/unit/data/clean/json/illegal_marked_object.json
@@ -1,0 +1,27 @@
+{
+	"content": [
+		{
+			"type": "paragraph",
+			"content": [
+				{
+					"type": "atlas_object",
+					"attrs": { "atlas": "", "sprite": "item/cobblestone" },
+					"marks": [
+						{ "type": "bold" },
+						{ "type": "italic" },
+						{ "type": "obfuscated" }
+					]
+				},
+				{
+					"type": "player_object",
+					"attrs": { "player": { "name": "Cbble_" }, "hat": true },
+					"marks": [
+						{ "type": "bold" },
+						{ "type": "italic" },
+						{ "type": "obfuscated" }
+					]
+				}
+			]
+		}
+	]
+}
