+++ src/tests/unit/data/clean/snbt/base16_shadow_color.snbt
@@ -1,1 +1,1 @@
-["",{shadow_color:"#ff0000",text:"oooo spooky shadows"}]
+["",{shadow_color:"#ffff0000",text:"oooo spooky shadows"}]
