+++ vite.config.ts
@@ -15,7 +15,7 @@ export default defineConfig({
 					name: "client",
 					environment: "jsdom",
 					clearMocks: true,
-					include: ["src/tests/unit/*.svelte.{test,spec}.{js,ts}"],
+					include: ["src/tests/unit/**/*.svelte.{test,spec}.{js,ts}"],
 					exclude: ["src/lib/server/**"],
 					setupFiles: ["./vitest-setup-client.ts"],
 				},
@@ -25,7 +25,7 @@ export default defineConfig({
 				test: {
 					name: "server",
 					environment: "node",
-					include: ["src/tests/unit/*.{test,spec}.{js,ts}"],
+					include: ["src/tests/unit/**/*.{test,spec}.{js,ts}"],
 					exclude: ["src/**/*.svelte.{test,spec}.{js,ts}"],
 				},
 			},
